+++ src/main/java/com/github/dockerjava/api/command/AttachContainerCmd.java
@@ -2,6 +2,7 @@ package com.github.dockerjava.api.command;
 
 import java.io.InputStream;
 
+import com.github.dockerjava.api.DockerClient;
 import com.github.dockerjava.api.NotFoundException;
 
 /**
@@ -20,7 +21,7 @@ import com.github.dockerjava.api.NotFoundException;
  *            - true or false, if true, print timestamps for every log line.
  *            Defaults to false.
  */
-public interface AttachContainerCmd extends DockerCmd<InputStream>{
+public interface AttachContainerCmd extends DockerCmd<InputStream> {
 
 	public String getContainerId();
 
@@ -36,8 +37,17 @@ public interface AttachContainerCmd extends DockerCmd<InputStream>{
 
 	public AttachContainerCmd withContainerId(String containerId);
 
+	/**
+	 * See {@link #withFollowStream(boolean)}
+	 */
 	public AttachContainerCmd withFollowStream();
 
+	/**
+	 * Following the stream means the resulting {@link InputStream} returned by
+	 * {@link #exec()} reads infinitely. So a {@link InputStream#read()} MAY
+	 * BLOCK FOREVER as long as no data is streamed from the docker host to
+	 * {@link DockerClient}!
+	 */
 	public AttachContainerCmd withFollowStream(boolean followStream);
 
 	public AttachContainerCmd withTimestamps(boolean timestamps);
@@ -51,19 +61,21 @@ public interface AttachContainerCmd extends DockerCmd<InputStream>{
 	public AttachContainerCmd withStdErr(boolean stderr);
 
 	public AttachContainerCmd withLogs(boolean logs);
-	
+
 	public AttachContainerCmd withLogs();
 
 	/**
-	 * Its the responsibility of the caller to consume and/or close the {@link InputStream} to prevent
-	 * connection leaks.
+	 * Its the responsibility of the caller to consume and/or close the
+	 * {@link InputStream} to prevent connection leaks.
 	 * 
-	 * @throws NotFoundException No such container 
+	 * @throws NotFoundException
+	 *             No such container
 	 */
 	@Override
 	public InputStream exec() throws NotFoundException;
-	
-	public static interface Exec extends DockerCmdExec<AttachContainerCmd, InputStream> {
+
+	public static interface Exec extends
+			DockerCmdExec<AttachContainerCmd, InputStream> {
 	}
 
 }