+++ src/main/java/com/github/dockerjava/api/command/PullImageCmd.java
@@ -1,6 +1,7 @@
 package com.github.dockerjava.api.command;
 
 import com.github.dockerjava.api.model.AuthConfig;
+import com.github.dockerjava.core.command.EventStreamReader;
 
 import java.io.InputStream;
 
@@ -33,6 +34,8 @@ public interface PullImageCmd extends DockerCmd<InputStream>{
     /**
      * Its the responsibility of the caller to consume and/or close the {@link InputStream} to prevent
 	 * connection leaks.
+     *
+     * @see {@link EventStreamReader}
      */
     @Override
     public InputStream exec();