+++ src/main/java/com/github/dockerjava/jaxrs/DockerCmdExecFactoryImpl.java
@@ -4,6 +4,7 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 import java.io.IOException;
 import java.net.URI;
+import java.util.concurrent.TimeUnit;
 
 import com.github.dockerjava.api.command.*;
 
@@ -15,6 +16,7 @@ import javax.ws.rs.client.Client;
 import javax.ws.rs.client.ClientBuilder;
 import javax.ws.rs.client.WebTarget;
 
+import org.apache.http.client.config.RequestConfig;
 import org.apache.http.config.RegistryBuilder;
 import org.apache.http.conn.socket.ConnectionSocketFactory;
 import org.apache.http.conn.socket.PlainConnectionSocketFactory;
@@ -22,7 +24,12 @@ import org.apache.http.conn.ssl.SSLConnectionSocketFactory;
 import org.apache.http.impl.conn.PoolingHttpClientConnectionManager;
 import org.glassfish.jersey.CommonProperties;
 import org.glassfish.jersey.apache.connector.ApacheClientProperties;
-import org.glassfish.jersey.apache.connector.ApacheConnectorProvider;
+
+
+//import org.glassfish.jersey.apache.connector.ApacheConnectorProvider;
+// see https://github.com/docker-java/docker-java/issues/196
+import com.github.dockerjava.jaxrs.connector.ApacheConnectorProvider;
+
 import org.glassfish.jersey.client.ClientConfig;
 import org.glassfish.jersey.client.ClientProperties;
 
@@ -33,6 +40,7 @@ import com.github.dockerjava.core.util.FollowRedirectsFilter;
 import com.github.dockerjava.core.util.JsonClientFilter;
 import com.github.dockerjava.core.util.ResponseStatusExceptionFilter;
 import com.github.dockerjava.core.util.SelectiveLoggingFilter;
+
 
 public class DockerCmdExecFactoryImpl implements DockerCmdExecFactory {
 
@@ -67,6 +75,8 @@ public class DockerCmdExecFactoryImpl implements DockerCmdExecFactory {
 			clientConfig.property(ClientProperties.READ_TIMEOUT, readTimeout);
 		}
 
+		//clientConfig.property(ClientProperties.CONNECT_TIMEOUT, 10000);
+
 		URI originalUri = dockerClientConfig.getUri();
 
 		SSLContext sslContext = null;
@@ -81,7 +91,8 @@ public class DockerCmdExecFactoryImpl implements DockerCmdExecFactory {
 		}
 
 		PoolingHttpClientConnectionManager connManager = new PoolingHttpClientConnectionManager(
-				getSchemeRegistry(originalUri, sslContext));
+				getSchemeRegistry(originalUri, sslContext), null, null, null, 10, TimeUnit.SECONDS);
+
 
 		if (dockerClientConfig.getMaxTotalConnections() != null)
 			connManager
@@ -92,6 +103,9 @@ public class DockerCmdExecFactoryImpl implements DockerCmdExecFactory {
 
 		clientConfig.property(ApacheClientProperties.CONNECTION_MANAGER,
 				connManager);
+
+		clientConfig.property(ApacheClientProperties.REQUEST_CONFIG,
+                RequestConfig.custom().setConnectionRequestTimeout(10).build());
 
 		ClientBuilder clientBuilder = ClientBuilder.newBuilder().withConfig(
 				clientConfig);
@@ -311,6 +325,11 @@ public class DockerCmdExecFactoryImpl implements DockerCmdExecFactory {
 	@Override
 	public EventsCmd.Exec createEventsCmdExec() {
 		return new EventsCmdExec(getBaseResource());
+	}
+
+	@Override
+    public StatsCmd.Exec createStatsCmdExec() {
+	    return new StatsCmdExec(getBaseResource());
 	}
 
 	@Override