+++ src/main/java/com/github/dockerjava/jaxrs/EventsCmdExec.java
@@ -22,68 +22,88 @@ import com.github.dockerjava.api.command.EventsCmd;
 import com.github.dockerjava.api.model.Event;
 import com.github.dockerjava.jaxrs.util.WrappedResponseInputStream;
 
-public class EventsCmdExec extends AbstrDockerCmdExec<EventsCmd, ExecutorService> implements EventsCmd.Exec {
-    private static final Logger LOGGER = LoggerFactory.getLogger(EventsCmdExec.class);
-    
-    public EventsCmdExec(WebTarget baseResource) {
-        super(baseResource);
-    }
+public class EventsCmdExec extends
+		AbstrDockerCmdExec<EventsCmd, ExecutorService> implements
+		EventsCmd.Exec {
+	private static final Logger LOGGER = LoggerFactory
+			.getLogger(EventsCmdExec.class);
 
-    @Override
-    protected ExecutorService execute(EventsCmd command) {
-    	ExecutorService executorService = Executors.newSingleThreadExecutor();
-    	
-        WebTarget webResource = getBaseResource().path("/events")
-                .queryParam("since", command.getSince())
-                .queryParam("until", command.getUntil());
+	public EventsCmdExec(WebTarget baseResource) {
+		super(baseResource);
+	}
 
-        LOGGER.trace("GET: {}", webResource);
-        EventNotifier eventNotifier = EventNotifier.create(command.getEventCallback(), webResource);
-        executorService.submit(eventNotifier);
-        return executorService;
-    }
-    
-    private static class EventNotifier implements Callable<Void> {
-        private static final JsonFactory JSON_FACTORY = new JsonFactory();
-        private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
+	@Override
+	protected ExecutorService execute(EventsCmd command) {
+		ExecutorService executorService = Executors.newSingleThreadExecutor();
 
-        private final EventCallback eventCallback;
-        private final WebTarget webTarget;
+		WebTarget webResource = getBaseResource().path("/events")
+				.queryParam("since", command.getSince())
+				.queryParam("until", command.getUntil());
 
-        private EventNotifier(EventCallback eventCallback, WebTarget webTarget) {
-            this.eventCallback = eventCallback;
-            this.webTarget = webTarget;
-        }
+		LOGGER.trace("GET: {}", webResource);
+		EventNotifier eventNotifier = EventNotifier.create(
+				command.getEventCallback(), webResource);
+		executorService.submit(eventNotifier);
+		return executorService;
+	}
 
-        public static EventNotifier create(EventCallback eventCallback, WebTarget webTarget) {
-            checkNotNull(eventCallback, "An EventCallback must be provided");
-            checkNotNull(webTarget, "An WebTarget must be provided");
-            return new EventNotifier(eventCallback, webTarget);
-        }
+	private static class EventNotifier implements Callable<Void> {
+		private static final JsonFactory JSON_FACTORY = new JsonFactory();
+		private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
 
-        @Override
-        public Void call() throws Exception {
-            int numEvents=0;
-            Response response = null;
-            try {
-                response = webTarget.request().get(Response.class);
-                InputStream inputStream = new WrappedResponseInputStream(response);
-                JsonParser jp = JSON_FACTORY.createParser(inputStream);
-                while (jp.nextToken() != JsonToken.END_OBJECT && !jp.isClosed() && eventCallback.isReceiving()) {
-                    eventCallback.onEvent(OBJECT_MAPPER.readValue(jp, Event.class));
-                    numEvents++;
-                }
-            }
-            catch(Exception e) {
-                eventCallback.onException(e);
-            }
-            finally {
-                if (response != null) {
+		private final EventCallback eventCallback;
+		private final WebTarget webTarget;
+
+		private EventNotifier(EventCallback eventCallback, WebTarget webTarget) {
+			this.eventCallback = eventCallback;
+			this.webTarget = webTarget;
+		}
+
+		public static EventNotifier create(EventCallback eventCallback,
+				WebTarget webTarget) {
+			checkNotNull(eventCallback, "An EventCallback must be provided");
+			checkNotNull(webTarget, "An WebTarget must be provided");
+			return new EventNotifier(eventCallback, webTarget);
+		}
+
+		@Override
+		public Void call() throws Exception {
+			int numEvents = 0;
+			Response response = null;
+			try {
+				response = webTarget.request().get(Response.class);
+				InputStream inputStream = new WrappedResponseInputStream(
+						response);
+				JsonParser jp = JSON_FACTORY.createParser(inputStream);
+				// The following condition looks strange but jp.nextToken() will block until there is an
+				// event from the docker server or the connection is terminated.
+				// therefore we want to check before getting an event (to prevent a blocking operation
+				// and after the event to make sure that the eventCallback is still interested in getting notified.
+				while (eventCallback.isReceiving() &&
+				       jp.nextToken() != JsonToken.END_OBJECT && !jp.isClosed() &&
+				       eventCallback.isReceiving()) {
+					try {
+						eventCallback.onEvent(OBJECT_MAPPER.readValue(jp,
+								Event.class));
+					} catch (Exception e) {
+						eventCallback.onException(e);
+					}
+					numEvents++;
+				}
+			} catch (Exception e) {
+				eventCallback.onException(e);
+			} finally {
+			    if (response != null) {
                     response.close();
                 }
-            }
-            eventCallback.onCompletion(numEvents);
-            return null;
-        }
-    }
+				try {
+					eventCallback.onCompletion(numEvents);
+				} catch (Exception e) {
+					eventCallback.onException(e);
+				}
+			}
+
+			return null;
+		}
+	}
 }